backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.9.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pdfplumber==0.11.4
authlib==1.3.2
httpx==0.28.0
email-validator==2.2.0
python-dotenv==1.0.1
bcrypt==4.2.0
itsdangerous==2.2.0
sentence-transformers==3.1.1
pymupdf==1.24.11
python-docx==1.1.2
openai==1.52.0
google-genai==0.2.1
